Add "Scroll Top" button and include jQuery library
Added a "Scroll Top" button to the footer in `_Layout.cshtml` to
enhance user navigation by allowing users to quickly return to
the top of the page. The button is styled with appropriate
classes and includes an icon. Also included a reference to the
jQuery library to support any jQuery-dependent functionality.